deployment/requirements.txt
selenium==4.18.1
pandas==2.2.1
requests==2.31.0
python-dotenv==1.0.1
Flask==3.0.2
webdriver_manager==4.0.1
gunicorn==21.2.0 